Handle a single proxyjump bastion across multiple stacks.
In this change, if a previous cluster gateway IP address and user
are defined, they will be reused.  This enables us to use a single
bastion host for a multi-stack deployment (provided it is not needed
before the stack in which it is created).  It also makes it possible
for us to introduce an independent bastion that is not actually part of
the deployment.