Return cluster group info as output after string replacement
Each cluster group can include an `info` key, whose value is expected to
be a JSON object encoded as a string. We substitute "$PRIMARY_IPS" in
the JSON object by the list of primary IPs allocated to this group, and
return it as another key of the `group_data` output value.

In the cluster template, info from each group is put together in an
`info` output value, resulting in a list of JSON objects.